feat: move the "helpers" content from common to backend
