+++ halo2_backend/src/lib.rs
@@ -1,9 +1,8 @@
 pub mod arithmetic;
-mod helpers;
+pub mod helpers;
 pub mod plonk;
 pub mod poly;
 pub mod transcript;
 
 // Internal re-exports
 pub use halo2_common::multicore;
-pub use halo2_common::SerdeFormat;
