+++ halo2_backend/src/plonk/permutation.rs
@@ -4,9 +4,8 @@ use crate::{
     arithmetic::CurveAffine,
     helpers::{polynomial_slice_byte_length, read_polynomial_vec, write_polynomial_slice},
     poly::{Coeff, ExtendedLagrangeCoeff, LagrangeCoeff, Polynomial},
-    SerdeFormat,
 };
-use halo2_common::helpers::{SerdeCurveAffine, SerdePrimeField};
+use crate::helpers::{SerdeCurveAffine, SerdePrimeField, SerdeFormat};
 // TODO: Remove the renaming
 pub use halo2_middleware::permutation::ArgumentMid as Argument;
