+++ halo2_backend/src/poly.rs
@@ -3,8 +3,7 @@
 //! the committed polynomials at arbitrary points.
 
 use crate::arithmetic::parallelize;
-use crate::SerdeFormat;
-use halo2_common::helpers::SerdePrimeField;
+use crate::helpers::{SerdePrimeField, SerdeFormat};
 
 use group::ff::Field;
 use halo2_middleware::poly::Rotation;
